--- api/lib/infrastructure/repositories/participations-for-campaign-management-repository.js
@@ -1,41 +1,0 @@
-import { knex } from '../../../db/knex-database-connection.js';
-import { ParticipationForCampaignManagement } from '../../domain/models/ParticipationForCampaignManagement.js';
-import { fetchPage } from '../utils/knex-utils.js';
-
-const findPaginatedParticipationsForCampaignManagement = async function ({ campaignId, page }) {
-  const query = knex('campaign-participations')
-    .select({
-      id: 'campaign-participations.id',
-      lastName: 'view-active-organization-learners.lastName',
-      firstName: 'view-active-organization-learners.firstName',
-      userId: 'users.id',
-      userFirstName: 'users.firstName',
-      userLastName: 'users.lastName',
-      participantExternalId: 'campaign-participations.participantExternalId',
-      status: 'campaign-participations.status',
-      createdAt: 'campaign-participations.createdAt',
-      sharedAt: 'campaign-participations.sharedAt',
-      deletedAt: 'campaign-participations.deletedAt',
-      deletedBy: 'deletedByUsers.id',
-      deletedByFirstName: 'deletedByUsers.firstName',
-      deletedByLastName: 'deletedByUsers.lastName',
-    })
-    .join(
-      'view-active-organization-learners',
-      'view-active-organization-learners.id',
-      'campaign-participations.organizationLearnerId',
-    )
-    .leftJoin('users as deletedByUsers', 'deletedByUsers.id', 'campaign-participations.deletedBy')
-    .innerJoin('users', 'users.id', 'campaign-participations.userId')
-    .where('campaignId', campaignId)
-    .orderBy(['lastName', 'firstName'], ['asc', 'asc']);
-
-  const { results, pagination } = await fetchPage(query, page);
-
-  const participationsForCampaignManagement = results.map(
-    (attributes) => new ParticipationForCampaignManagement(attributes),
-  );
-  return { models: participationsForCampaignManagement, meta: { ...pagination } };
-};
-
-export { findPaginatedParticipationsForCampaignManagement };
