+++ api/src/prescription/campaign-participation/domain/models/ParticipationForCampaignManagement.js
