+++ api/src/prescription/campaign-participation/domain/usecases/find-paginated-participations-for-campaign-management.js
