+++ api/src/prescription/campaign-participation/infrastructure/serializers/jsonapi/participation-for-campaign-management-serializer.js
