+++ api/src/prescription/campaign/domain/usecases/get-campaign-details-management.js
