+++ api/src/prescription/campaign/infrastructure/serializers/jsonapi/campaign-details-management-serializer.js
