+++ api/tests/prescription/campaign-participation/unit/application/campaign-participation-route_test.js
@@ -179,4 +179,49 @@ describe('Unit | Application | Router | campaign-participation-router ', functio
       expect(response.statusCode).to.equal(403);
     });
   });
+
+  describe('GET /api/admin/campaigns/{id}/participations', function () {
+    it('should return 200', async function () {
+      // given
+      sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').returns(() => true);
+      sinon
+        .stub(campaignParticipationController, 'findPaginatedParticipationsForCampaignManagement')
+        .callsFake((request, h) => h.response('ok').code(200));
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/admin/campaigns/1/participations');
+
+      // then
+      expect(response.statusCode).to.equal(200);
+    });
+
+    it('should return 400 with an invalid campaign id', async function () {
+      // given
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/admin/campaigns/invalid/participations');
+
+      // then
+      expect(response.statusCode).to.equal(400);
+    });
+
+    it('should return 403 when unauthorized', async function () {
+      // given
+      sinon
+        .stub(securityPreHandlers, 'hasAtLeastOneAccessOf')
+        .returns((request, h) => h.response().code(403).takeover());
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/admin/campaigns/1/participations');
+
+      // then
+      expect(response.statusCode).to.equal(403);
+    });
+  });
 });
