+++ api/tests/prescription/campaign-participation/unit/domain/models/ParticipationForCampaignManagement_test.js
@@ -1,5 +1,5 @@
-import { ParticipationForCampaignManagement } from '../../../../lib/domain/models/ParticipationForCampaignManagement.js';
-import { expect } from '../../../test-helper.js';
+import { ParticipationForCampaignManagement } from '../../../../../../src/prescription/campaign-participation/domain/models/ParticipationForCampaignManagement.js';
+import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | ParticipationForCampaignManagement', function () {
   context('#userFullName', function () {
