+++ api/tests/prescription/campaign-participation/unit/infrastructure/serializers/jsonapi/participation-for-campaign-management-serializer_test.js
@@ -1,5 +1,5 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/participation-for-campaign-management-serializer.js';
-import { domainBuilder, expect } from '../../../../test-helper.js';
+import * as serializer from '../../../../../../../src/prescription/campaign-participation/infrastructure/serializers/jsonapi/participation-for-campaign-management-serializer.js';
+import { domainBuilder, expect } from '../../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | participation-for-campaign-management-serializer', function () {
   describe('#serialize()', function () {
