+++ api/tests/prescription/campaign/unit/application/campaign-detail-route_test.js
@@ -235,4 +235,47 @@ describe('Unit | Application | Router | campaign-detail-router ', function () {
       expect(result.statusCode).to.equal(400);
     });
   });
+
+  describe('GET /api/admin/campaigns/{id}', function () {
+    it('should return 200', async function () {
+      // given
+      sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').returns(() => true);
+      sinon.stub(campaignDetailController, 'getCampaignDetails').callsFake((request, h) => h.response('ok').code(200));
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/admin/campaigns/1');
+
+      // then
+      expect(response.statusCode).to.equal(200);
+    });
+
+    it('should return 400 with an invalid campaign id', async function () {
+      // given
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/admin/campaigns/invalid');
+
+      // then
+      expect(response.statusCode).to.equal(400);
+    });
+
+    it('should return 403 when unauthorized', async function () {
+      // given
+      sinon
+        .stub(securityPreHandlers, 'hasAtLeastOneAccessOf')
+        .returns((request, h) => h.response().code(403).takeover());
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/admin/campaigns/1');
+
+      // then
+      expect(response.statusCode).to.equal(403);
+    });
+  });
 });
