+++ api/tests/prescription/campaign/unit/infrastrucutre/serializers/jsonapi/campaign-details-management-serializer_test.js
@@ -1,5 +1,5 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/campaign-details-management-serializer.js';
-import { expect } from '../../../../test-helper.js';
+import * as serializer from '../../../../../../../src/prescription/campaign/infrastructure/serializers/jsonapi/campaign-details-management-serializer.js';
+import { expect } from '../../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | campaign-details-management-serializer', function () {
   describe('#serialize()', function () {
