+++ api/tests/tooling/domain-builder/factory/build-participation-for-campaign-management.js
@@ -1,4 +1,4 @@
-import { ParticipationForCampaignManagement } from '../../../../lib/domain/models/ParticipationForCampaignManagement.js';
+import { ParticipationForCampaignManagement } from '../../../../src/prescription/campaign-participation/domain/models/ParticipationForCampaignManagement.js';
 import { CampaignParticipationStatuses } from '../../../../src/prescription/shared/domain/constants.js';
 
 const buildParticipationForCampaignManagement = function ({
